# target/classes/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Oderstore - OpenAPI 3.0
  description: |-
    This is a sample Oder Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the oder store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Oderstore, then click [here](https://editor.swagger.io/?url=https://oderstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Oderstore OAS 2.0` menu option!_
    
    Some useful links:
    - [The Oder Store repository](https://github.com/swagger-api/swagger-oderstore)
    - [The source API definition for the Oder Store](https://github.com/swagger-api/swagger-oderstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://oderstore3.swagger.io/api/v3
...
paths:
  /order:
    put:
      tags:
        - order
      summary: Update an existing order
      description: Update an existing order by Id
      operationId: orderOder
      requestBody:
        description: Update an existent order in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
          application/xml:
            schema:
              $ref: '#/components/schemas/Order'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Order'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid ID supplied
        '404':
          description: Order not found
        '422':
          description: Validation exception
      security:
        - orderstore_auth:
            - write:oders
            - read:oders
    post:
      tags:
        - order
      summary: Add a new order
      description: Add a new order
      operationId: addOrder
      requestBody:
        description: Create a Order
        x-constraints: "@ValidCake"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
          application/xml:
            schema:
              $ref: '#/components/schemas/Order'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Order'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
      security:
        - orderstore_auth:
            - write:oders
            - read:oders
...
components:
  schemas:
    Order:
      type: object
      required:
        - Cakes
        - PickupPoint
      properties:
        id:
          type: integer
          format: int64
          example: 10
        oderId:
          type: integer
          format: int64
          example: 198772
        PickUpPoint:
          $ref: 'components/schemas/PickUpPoint'
        Cakes:
          type: array
          items:
            $ref: 'components/schemas/Cake'
        shipDate:
          type: string
          format: date-time
        status:
          type: string
          description: Order Status
          example: approved
          enum:
            - placed
            - approved
            - delivered
        complete:
          type: boolean
      xml:
        name: order
...
  securitySchemes:
    orderstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://oderstore3.swagger.io/oauth/authorize
          scopes:
            write:oders: modify oders in your account
            read:oders: read your oders
    api_key:
      type: apiKey
      name: api_key
      in: header